feat: FireStoreManager와 FirebaseManager 분리
